Data.ByteString.Lazy.UTF8

Description

Documentation

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A lazy ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Lazy.Char8 it can be interpreted as containing 8-bit characters.

decode :: ByteString -> Maybe (Char, Int64)

Try to extract a character from a byte string. Returns Nothing if there are no more bytes in the byte string. Otherwise, it returns a decoded character and the number of bytes used in its representation. Errors are replaced by character '\0xFFFD'.

replacement_char :: Char

This character is used to mark errors in a UTF8 encoded string.

uncons :: ByteString -> Maybe (Char, ByteString)

Get the first character of a byte string, if any. Malformed characters are replaced by '\0xFFFD'.

splitAt :: Int64 -> ByteString -> (ByteString, ByteString)

Split after a given number of characters. Negative values are treated as if they are 0.

take :: Int64 -> ByteString -> ByteString

take n s returns the first n characters of s. If s has less than n characters, then we return the whole of s.

drop :: Int64 -> ByteString -> ByteString

drop n s returns the s without its first n characters. If s has less than n characters, then we return an empty string.

span :: (Char -> Bool) -> ByteString -> (ByteString, ByteString)

Split a string into two parts: the first is the longest prefix that contains only characters that satisfy the predicate; the second part is the rest of the string. Invalid characters are passed as '\0xFFFD' to the predicate.

break :: (Char -> Bool) -> ByteString -> (ByteString, ByteString)

Split a string into two parts: the first is the longest prefix that contains only characters that do not satisfy the predicate; the second part is the rest of the string. Invalid characters are passed as '\0xFFFD' to the predicate.

fromString :: String -> ByteString

Converts a Haskell string into a UTF8 encoded bytestring.

toString :: ByteString -> String

Convert a UTF8 encoded bytestring into a Haskell string. Invalid characters are replaced with '\0xFFFD'.

foldl :: (a -> Char -> a) -> a -> ByteString -> a

Traverse a bytestring (left biased). This function is strict in the accumulator.

foldr :: (Char -> a -> a) -> a -> ByteString -> a

Traverse a bytestring (right biased).

length :: ByteString -> Int

Counts the number of characters encoded in the bytestring. Note that this includes replacement characters.

lines :: ByteString -> [ByteString]

Split a string into a list of lines. Lines are terminated by '\n' or the end of the string. Empty lines may not be terminated by the end of the string. See also lines'.

lines' :: ByteString -> [ByteString]

Split a string into a list of lines. Lines are terminated by '\n' or the end of the string. Empty lines may not be terminated by the end of the string. This function preserves the terminators. See also lines.